Move getSockets to new file
